feat:outsourced likeButton
